% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\docType{data}
\name{conv}
\alias{conv}
\title{Data frame of coral size conversion factors}
\format{A data frame:
\describe{
  \item{\code{spec}}{chr string of species name}
  \item{\code{conv}}{numeric for conversion factor}
}}
\usage{
conv
}
\description{
Morphological conversion factors are based on Table J-1 in EPA/600/R-13/350 (same as Table 3-2 in EPA/600/R-12/029) and the equation
}
\keyword{datasets}
