% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acrorb.R
\name{get_acrorb}
\alias{get_acrorb}
\title{Get dominance of acropora/orbicella genera}
\usage{
get_acrorb(dat_in, gen = c("Acropora", "Orbicella"))
}
\arguments{
\item{dat_in}{coral demographic data, does not need estimated surface area}

\item{gen}{chr string of genera for sensitive/rare species}
}
\value{
data frame of relative abundance and csa of the genera for all stations
}
\description{
Get dominance of acropora/orbicella genera
}

