% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rel_abu.R
\name{get_rel_abu}
\alias{get_rel_abu}
\title{Coral relative abundance}
\usage{
get_rel_abu(dat_in)
}
\arguments{
\item{dat_in}{raw coral demographic data}
}
\value{
data frame of species abundance and relative abundance by station
}
\description{
Coral relative abundance
}
\details{
abundance is simply a count of species at a station
}

