% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sick.R
\name{get_sick}
\alias{get_sick}
\title{Get proportion of individuals at a station that are diseased or bleached}
\usage{
get_sick(dat_in)
}
\arguments{
\item{dat_in}{coral demographic data, does not need estimated surface area}
}
\value{
data frame of bleached or diseased proportion at each station
}
\description{
Get proportion of individuals at a station that are diseased or bleached
}
\examples{
get_sick(crl_dem)
}

