% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sr_div.R
\name{get_sr_div}
\alias{get_sr_div}
\title{Diversity of sensitive and rare species}
\usage{
get_sr_div(dat_in, gen = c("Eusmilia", "Isophyllastrea", "Isophyllia",
  "Mycetophyllia", "Scolymia"))
}
\arguments{
\item{dat_in}{coral demographic data, does not need estimated surface area}

\item{gen}{chr string of genera for sensitive/rare species}
}
\value{
data frame with one row per station, richness of sens/rare species, relative abundance of sense/rare species, and estimated diversity
}
\description{
Diversity of sensitive and rare species
}
\details{
diversity is estimated with \code{\link[vegan]{diversity}}
}

