% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stony_mets.R
\name{get_stony_mets}
\alias{get_stony_mets}
\title{Get stony coral metrics}
\usage{
get_stony_mets(dat_in, rems = "No Coral Observed")
}
\arguments{
\item{dat_in}{raw coral demographic data}

\item{rems}{chr string of species to remove, used to remove stations with no coral}
}
\value{
Data frame of coral metrics with one row per station
}
\description{
Get all stony coral metrics
}
\examples{
get_stony_mets(crl_dem)
}

