% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tot_div.R
\name{get_tot_div}
\alias{get_tot_div}
\title{Get total alpha diversity}
\usage{
get_tot_div(dat_in)
}
\arguments{
\item{dat_in}{coral demographic data, does not need estimated surface area}
}
\value{
data frame with one row per station, total richness, and estimated diversity for all species
}
\description{
Get total alpha diversity
}

