% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzz.R
\name{plot_fuzz}
\alias{plot_fuzz}
\title{Plot fuzzy metrics}
\usage{
plot_fuzz(met_in, scr_in, met, cols = "Paired", met_val = NULL,
  sdchg = NULL, widths = c(0.6, 1, 0.6, 0.6))
}
\arguments{
\item{met_in}{input \code{data.frame} of coral metrics}

\item{scr_in}{input \code{data.frame} of site BCG scores, see details}

\item{met}{chr string of metric to plot}

\item{cols}{chr string of palette for plot colors, from \code{\link[RColorBrewer]{brewer.pal}}}

\item{met_val}{numeric value of metric to show on plot}

\item{sdchg}{optional list passed to \code{\link{met_crvs}} for manually changing standard deviation estimates}

\item{widths}{numeric vector indicating widths of each of four plots, passed to \code{\link[gridExtra]{grid.arrange}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot fuzzy metrics given based on expert BCG levels
}
\details{
The input \code{scr_in} data has two columns labelled \code{station_code} and \code{scr}. The station codes should match those in \code{met_in}. The \code{scr} values for each station can be numeric or character string BCG levels that represent qualitative rankings.
}
\examples{
\dontrun{
# coral metrics
met_in <- get_stony_mets(crl_dem)

# bcg scores for each station
station_code <- c(1:5)
scr <- c(2, 5, 3, 2, 4)
scr_in <- data.frame(station_code, scr) 

# plot
plot_fuzz(met_in, scr_in, met = 'tot_rich')

# plot, manual change of standard deviation
sdchg <- list(tot_rich = c(2, 2, 2, 2))
plot_fuzz(met_in, scr_in, met = 'tot_rich', sdchg = sdchg)
}
}
\seealso{
\code{\link{met_crvs}}, \code{\link{met_dist}}
}
