% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acrorb.R
\name{get_acrorb}
\alias{get_acrorb}
\title{Get dominance of acropora/orbicella genera}
\usage{
get_acrorb(dat_in, gen = c("Acropora", "Orbicella"))
}
\arguments{
\item{dat_in}{coral demographic data with estimated csa (from \code{\link{est_3d}})}

\item{gen}{chr string of genera for sensitive/rare species}
}
\value{
data frame of relative abundance and csa of the genera for all stations
}
\description{
Get dominance of acropora/orbicella genera
}
\examples{
crl_dem$csa <- with(crl_dem, est_3d(species_name, Height, MaxDiam, PerpDiam))
get_acrorb(crl_dem)
}
