% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_sz.R
\name{get_col_sz}
\alias{get_col_sz}
\title{Get colony size distributions}
\usage{
get_col_sz(dat_in)
}
\arguments{
\item{dat_in}{coral demographic data with esimated csa per species (from \code{\link{est_3d}})}
}
\value{
data frame of station and colonies with indicator if recruits present and total csa (cm2)
}
\description{
Get colony size distributions
}
\examples{
crl_dem$csa <- with(crl_dem, est_3d(species_name, Height, MaxDiam, PerpDiam))
get_col_sz(crl_dem)
}

