% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lrg_mort.R
\name{get_lrg_mort}
\alias{get_lrg_mort}
\title{Get mortality of large reef building genera}
\usage{
get_lrg_mort(dat_in, gen = c("Acropora", "Colpophyllia", "Dendrogyra",
  "Orbicella", "Pseudodiploria"))
}
\arguments{
\item{dat_in}{coral demographic data with esimated csa (from \code{\link{est_3d}})}

\item{gen}{chr string of large genera}
}
\value{
data frame with counts of large genera (abundance as number of individuals/colonies) and one estimate of mortality per station estimated as average of individual mortality from 0 - 1
}
\description{
Get mortality of large reef building genera
}
\examples{
crl_dem$csa <- with(crl_dem, est_3d(species_name, Height, MaxDiam, PerpDiam))
get_lrg_mort(crl_dem)
}
