% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_crvs.R
\name{met_crvs}
\alias{met_crvs}
\title{Get probability distribution curves}
\usage{
met_crvs(met_in, scr_in, n = 1000)
}
\arguments{
\item{met_in}{input \code{data.frame} of coral metrics}

\item{scr_in}{input \code{data.frame} of site BCG scores, see details}

\item{n}{numeric indicating number of values across the range for each metric to predict the density curves}
}
\value{
A two-element list named \code{met_in} and \code{crvs}, where the former is the joined input data with BCG scores in \code{scr_in} and the latter is estimated density curves from the raw distributions of each.
}
\description{
Get probability distribution curves from raw metrics for plotting
}
\details{
The input \code{scr_in} data has two columns labelled \code{station_code} and \code{scr}. The station codes should match those in \code{met_in}. The \code{scr} values for each station can be numeric or character string BCG levels that represent qualitative rankings. 

The density curves are based on maximum-likelihood estimates of the mean and standard deviation for a normal curve corresponding to the raw metric data at each BCG level.
}
\examples{
\dontrun{
# coral metrics
met_in <- get_stony_mets(crl_dem)

# bcg scores for each station
station_code <- c(1:5)
scr <- c(2, 5, 3, 2, 4)
scr_in <- data.frame(station_code, scr) 

# get curves
met_crvs(met_in, scr_in)
} 
}
\seealso{
\code{\link{plot_fuzz}}
}

